+++ test/unit.test.ts
@@ -7,7 +7,6 @@ import './unit/instance-of-check.test.js';
 import './unit/rx-schema.test.js';
 import './unit/key-compression.test.js';
 import './unit/bug-report.test.js';
-import './unit/replication-graphql.test.js';  // TODO move down under replication
 import './unit/rx-database.test.js';
 import './unit/cache-replacement-policy.test';
 import './unit/rx-collection.test.js';
@@ -29,6 +28,7 @@ import './unit/data-migration.test.js';
 import './unit/backup.test.js';
 import './unit/leader-election.test.js';
 import './unit/replication.test.js';
+import './unit/replication-graphql.test.js';
 import './unit/encryption.test.js';
 import './unit/import-export.test.js';
 import './unit/cross-instance.test.js';